MutexGuard should not implement Send
If MutexGuard implements send it violates the soundness of the mutex
lock by giving access to the inner T from a thread that did not acquire
the lock. That is to say the following snippet should not compile, but
currently does:

use spin::Mutex;

fn main() {
    let mutex = Mutex::new(0usize);
    let mut locked = mutex.lock();

    crossbeam::thread::scope(move |s| {
        s.spawn(move |_| {
            *locked += 1;
            println!("locked: {}", *locked);
        });
    }).unwrap();
}

This changes the MutexGuard to store a *mut T rather than an &mut T
deferring the creation of an &mut T to the Deref impl. Since *mut T is
!Send that makes the MutexGuard also !Send.