once: merge two tests to prevent flaky test failures
Without this change, one of these two test methods fails about .75% of
the time (3 times in 500 runs). With this change, all tests pass in
5000 out of 5000 runs, so flaky failures should be a thing of the
past. I tried to reason a way to use a mutex or something here to
avoid fusing the methods, but I couldn't reason out a way to make it
work gracefully so I took the easy way out.